requirements.txt
yfinance==0.2.18
matplotlib==3.7.1
pandas==2.0.1
reportlab==3.6.12
python-dotenv==1.0.0
requests==2.30.0
numpy==1.24.3